[IFT] Begin reworking the patch application api.
Prepare the api to work within the PatchApplicationGroup framework.
